refactor: Move named pipe connection class to its own file and out of the NamedPipeServer class
