+++ src/CoreHook.IPC/NamedPipes/NamedPipeServer.cs
@@ -154,65 +154,5 @@ namespace CoreHook.IPC.NamedPipes
             Console.WriteLine(message);
             Console.WriteLine(e);
         }
-
-        public class Connection : IConnection
-        {
-            public NamedPipeServerStream ServerStream { get; }
-
-            private readonly StreamReader _reader;
-            private readonly StreamWriter _writer;
-            private readonly Func<bool> _isStopping;
-
-            public Connection(NamedPipeServerStream serverStream, Func<bool> isStopping)
-            {
-                ServerStream = serverStream;
-
-                _isStopping = isStopping;
-                _reader = new StreamReader(ServerStream);
-                _writer = new StreamWriter(ServerStream);
-            }
-
-            public bool IsConnected
-            {
-                get { return !_isStopping() && ServerStream.IsConnected; }
-            }
-
-            public NamedPipeMessages.IMessage ReadMessage()
-            {
-                return NamedPipeMessages.Message.FromString(ReadRequest());
-            }
-
-            public string ReadRequest()
-            {
-                try
-                {
-                    return _reader.ReadLine();
-                }
-                catch (IOException)
-                {
-                    return null;
-                }
-            }
-
-            public bool TrySendResponse(string message)
-            {
-                try
-                {
-                    _writer.WriteLine(message);
-                    _writer.Flush();
-
-                    return true;
-                }
-                catch (IOException)
-                {
-                    return false;
-                }
-            }
-
-            public bool TrySendResponse(NamedPipeMessages.Message message)
-            {
-                return TrySendResponse(message.ToString());
-            }
-        }
     }
 }
